negotiation/alphaNego-IJCAI22/conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.3.0=mkl
absl-py=0.4.1=py35_0
astor=0.7.1=py35_0
backcall=0.1.0=pypi_0
blas=1.0=mkl
ca-certificates=2021.10.26=h06a4308_2
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.11.5=py35he75722e_1
chardet=3.0.4=pypi_0
click=7.1.1=pypi_0
cocoa=0.1=dev_0
cudatoolkit=10.0.130=0
cycler=0.10.0=pypi_0
datasketch=1.2.3=pypi_0
decorator=4.4.2=pypi_0
flask=0.12.2=pypi_0
flask-socketio=2.8.5=pypi_0
freetype=2.9.1=h8a8886c_1
gast=0.4.0=py_0
grpcio=1.12.1=py35hdbcaa40_0
idna=2.9=pypi_0
intel-openmp=2020.0=166
ipython=7.9.0=pypi_0
ipython-genutils=0.2.0=pypi_0
itsdangerous=1.1.0=pypi_0
jedi=0.16.0=pypi_0
jinja2=2.11.1=pypi_0
joblib=0.14.1=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.1.0=pypi_0
langdetect=1.0.7=pypi_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.6.0=hdbcaa40_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
markdown=2.6.11=py35_0
markupsafe=1.1.1=pypi_0
matplotlib=3.0.3=pypi_0
mkl=2020.0=166
ncurses=6.2=he6710b0_0
ninja=1.8.2=py35h6bb024c_1
nltk=3.4.5=pypi_0
numpy=1.14.2=py35hdbf6ddf_0
olefile=0.46=py35_0
openssl=1.0.2u=h7b6447c_0
panda=0.3.1=pypi_0
pandas=0.25.3=pypi_0
parso=0.6.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=5.2.0=py35heded4f4_0
pip=10.0.1=py35_0
prompt-toolkit=2.0.10=pypi_0
protobuf=3.11.3=pypi_0
ptyprocess=0.6.0=pypi_0
pycparser=2.19=py35_0
pygments=2.6.1=pypi_0
pyparsing=2.4.6=pypi_0
python=3.5.6=hc3d631a_0
python-dateutil=2.8.1=pypi_0
python-engineio=3.11.2=pypi_0
python-socketio=4.4.0=pypi_0
pytorch=1.1.0=py3.5_cuda10.0.130_cudnn7.5.1_0
pytz=2019.3=pypi_0
readline=7.0=h7b6447c_5
redis=3.4.1=pypi_0
requests=2.23.0=pypi_0
scikit-learn=0.22.2.post1=pypi_0
scipy=1.4.1=pypi_0
seaborn=0.9.0=pypi_0
sentencepiece=0.1.85=pypi_0
setuptools=40.2.0=py35_0
six=1.11.0=py35_1
sqlite=3.31.1=h7b6447c_0
tensorboard=1.10.0=py35hf484d3e_0
tensorboardx=2.1=pypi_0
tensorflow=1.10.0=mkl_py35heddcb22_0
tensorflow-base=1.10.0=mkl_py35h3c3e929_0
termcolor=1.1.0=py35_1
tk=8.6.8=hbc83047_0
torchtext=0.5.0=pypi_0
torchvision=0.3.0=py35_cu10.0.130_1
tqdm=4.43.0=pypi_0
traitlets=4.3.3=pypi_0
ujson=1.35=pypi_0
urllib3=1.25.8=pypi_0
wcwidth=0.1.8=pypi_0
werkzeug=1.0.0=pypi_0
wheel=0.31.1=py35_0
whoosh=2.7.4=pypi_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
